backend/requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
firebase-admin==4.5.3
Flask==2.3.3
flask-firebase-admin==0.2.4
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
google-api-core==1.34.1
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.48.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
jws==0.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
oauth2client==4.1.3
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-jwt==4.1.0
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
uWSGI==2.0.25.1
Werkzeug==3.0.3
